refactor: extract post module
