+++ src/builder.rs
@@ -1,4 +1,5 @@
 mod homepage;
+mod post;
 
 use crate::parser;
 use std::{
@@ -28,24 +29,6 @@ fn create_build_dir() -> () {
     }
 }
 
-fn get_permalink_from_title(post_title: String) -> String {
-    let lowercase = post_title.to_lowercase();
-    return str::replace(&lowercase, " ", "-");
-}
-
-fn get_post_full_path(post: &parser::Post) -> String {
-    let build_path = get_build_dir();
-    let post_dir_name = get_permalink_from_title(String::from(&post.frontmatter.title));
-    let full_dir_path = format!("{}/{}", build_path, post_dir_name);
-    return full_dir_path;
-}
-
-fn create_post_file(post: &parser::Post) -> () {
-    let full_path = get_post_full_path(&post);
-    let file_path = format!("{}/index.html", full_path);
-    create_file(&post.html, &file_path)
-}
-
 fn create_file(contents: &str, file_path: &str) -> () {
     let mut file = File::create(file_path).unwrap();
     match write!(file, "{}", contents) {
@@ -56,21 +39,6 @@ fn create_file(contents: &str, file_path: &str) -> () {
             println!("failed to create file, see error: {}", msg);
         }
     }
-}
-
-fn create_post_dir_and_file(post: &parser::Post) -> () {
-    let full_dir_path = get_post_full_path(&post);
-
-    match fs::create_dir(full_dir_path) {
-        Ok(_) => {
-            // println!("created dir succesfully");
-        }
-        Err(msg) => {
-            println!("failed to create directory, see error: {}", msg);
-        }
-    }
-
-    create_post_file(post);
 }
 
 fn add_public_assets_to_build() -> () {
@@ -89,7 +57,7 @@ pub fn build(posts: &Vec<parser::Post>) -> () {
     create_build_dir();
 
     for post in posts {
-        create_post_dir_and_file(&post);
+        post::create_post_dir_and_file(&post);
     }
 
     homepage::create_homepage(&posts);
